requirements.txt
fastapi==0.115.0
google-genai==0.3.0
uvicorn==0.30.6
